Rename local-config.tsx in remix app
