+++ dev-projects/remix/app/routes/api.keystatic.$.tsx
@@ -1,5 +1,5 @@
 import { LoaderFunction } from '@remix-run/node';
 import { handleLoader } from '@keystatic/remix/api';
-import config from '../../local-config';
+import config from '../../keystatic.config';
 
 export const loader: LoaderFunction = args => handleLoader({ config }, args);
