+++ dev-projects/remix/app/routes/keystatic.$.tsx
@@ -1,4 +1,4 @@
 import { makePage } from '@keystatic/remix/ui';
-import config from '../../local-config';
+import config from '../../keystatic.config';
 
 export default makePage(config);
